* use constructor injection * also add the default tree whatever it is but it seems to have a co2 saving bound * inject also co2 bindings * inject planted on date for trees hopefully the 1970-01-01 * add co2 calculation in co2 repository for all users and trees * change tree planted on to timestamp * change co2 saving to double like in original db

+++ src/main/java/org/dicadeveloper/weplantaforest/reports/co2/Co2Dto.java
@@ -8,9 +8,9 @@ public class Co2Dto {
     private long _treesCount;
 
     @Mapping("co2")
-    private float _co2;
+    private double _co2;
 
-    public Co2Dto(long treesCount, float co2) {
+    public Co2Dto(long treesCount, double co2) {
         _treesCount = treesCount;
         _co2 = co2;
     }
@@ -23,7 +23,7 @@ public class Co2Dto {
         _treesCount = trees;
     }
 
-    public float getCo2() {
+    public double getCo2() {
         return _co2;
     }
